pyproject.toml
[tool.poetry]
name = "vybe-ai-assistant"
version = "0.8.0"
description = "Local Agentic AI Assistant with RAG, Code Interpretation, and Multi-Modal Capabilities"
authors = ["Vybe Team <support@vybe.local>"]
readme = "README.md"
packages = [{include = "vybe_app"}]

[tool.poetry.dependencies]
python = "^3.11"

# Core Web Framework
Flask = "3.0.3"
Flask-SQLAlchemy = "3.1.1"
Flask-SocketIO = "5.3.6"
Flask-Login = "0.6.3"
Flask-WTF = "1.2.1"
Flask-CORS = "4.0.0"
Werkzeug = "3.0.3"
eventlet = "0.36.1"
gunicorn = "^21.2.0"
click = "^8.0.0"

# Database
SQLAlchemy = "2.0.31"

# Security & Authentication
bcrypt = "4.1.3"
Flask-Limiter = "3.5.0"
bleach = "6.1.0"
cryptography = "42.0.7"

# RAG & Vector Database
chromadb = "0.5.3"

# LLM Inference with Large Context Windows
llama-cpp-python = {version = "^0.3.15", extras = ["server"]}

# Scientific Computing
numpy = "^1.24.0"
Pillow = "10.4.0"
requests = "2.32.3"
beautifulsoup4 = "4.12.3"
lxml = "^4.9.0"

# Web Scraping & Document Processing
playwright = "1.44.0"
trafilatura = "^1.6.0"
PyPDF2 = "3.0.1"
aiohttp = "3.9.5"

# Data Science & Visualization - Python 3.13 compatible
matplotlib = "^3.8.0"
pandas = "^2.1.0"
scipy = "^1.11.0"
seaborn = "^0.12.0"
plotly = "^5.17.0"

# Background Tasks & System Monitoring
APScheduler = "3.10.4"
psutil = "5.9.8"
pynvml = "11.5.0"

# Utilities
colorama = "0.4.6"
tqdm = "4.66.4"
python-dotenv = "1.0.1"
markdown = "3.6"
cairosvg = "2.7.1"

# Audio Processing & TTS
edge-tts = "^6.1.0"
ffmpeg-python = "^0.2.0"
pydub = "^0.25.1"

# Computer Vision & Additional ML
opencv-python = "^4.8.0"
scikit-learn = "^1.3.0"
numba = "^0.58.0"

# Home Assistant Integration
homeassistant-api = "5.0.0"
zeroconf = "0.132.2"

# UI Components
customtkinter = "5.2.2"

# ONNX Runtime - Python 3.13 compatible
onnxruntime = "^1.20.0"

[tool.poetry.group.ai.dependencies]
# AI/ML Dependencies - Optional group (Python 3.11-3.12 only due to compatibility)
torch = {version = "^2.0.0", python = "^3.11,<3.13"}
transformers = {version = "^4.35.0", python = "^3.11,<3.13"}
accelerate = {version = "^0.20.0", python = "^3.11,<3.13"}
diffusers = {version = "^0.25.0", python = "^3.11,<3.13"}
safetensors = {version = "^0.4.0", python = "^3.11,<3.13"}

[tool.poetry.group.audio.dependencies]
# Audio/Video Processing - Optional group
soundfile = {version = "^0.12.1", python = "^3.11,<3.13"}
librosa = {version = "^0.10.0", python = "^3.11,<3.13"}
moviepy = {version = "^1.0.3", python = "^3.11,<3.13"}

[tool.poetry.group.dev.dependencies]
# Development Dependencies
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vybe = "run:main"

# Ruff configuration (preserved from original)
[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
extend-select = ["B", "C", "I", "W"]
exclude = ["extensions", "extensions-disabled", "vybe-env"]
ignore = [
    "E501", # Line too long
    "E721", # Do not compare types, use `isinstance`
    "E731", # Do not assign a `lambda` expression, use a `def`
    "I001", # Import block is un-sorted or un-formatted
    "C901", # Function is too complex
    "C408", # Rewrite as a literal
    "W605", # invalid escape sequence
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.poetry.urls]
"Homepage" = "https://github.com/socalcium/Vybe-Local-Agentic-Container"
"Bug Tracker" = "https://github.com/socalcium/Vybe-Local-Agentic-Container/issues"
